Extract UseSegmentPositions type
