+++ tooling/lsp/src/lib.rs
@@ -70,6 +70,7 @@ mod solver;
 mod tests;
 mod trait_impl_method_stub_generator;
 mod types;
+mod use_segment_positions;
 mod utils;
 mod visibility;
